Add templates config file
