+++ src/docs/templates/types.ts
@@ -1,0 +1,14 @@
+export type Template = {
+  id: string;
+  name: string;
+  shortDescription: string;
+  description: string;
+  categories: TemplateCategory[];
+};
+
+type TemplateCategory = {
+  name: string;
+  description: string;
+  imageUrl: string[];
+  href: string;
+};
